centroid does now work (but needs facelifting). float type of lat and lon in struct osm_node cause fatal precision errors. now fixed, but needs additional 8 bytes for each node, yet

+++ osm_inplace.h
@@ -43,7 +43,7 @@ struct osm_node
 {
    // osm data
    int64_t id;
-   float lat, lon;
+   double lat, lon;
    int ver, cs, uid;
    int vis;
    time_t tim;
